+++ rollup.config.js
@@ -8,6 +8,9 @@ export default [
             file: "./dist/index.js",
             format: "cjs",
             sourcemap: true,
+            external: [
+                "hyperapp"
+            ]
         },
         plugins: [
             typescript({
@@ -15,6 +18,51 @@ export default [
                     declaration: true,
                 },
             }),
+        ]
+    },
+    {
+        input: "./src/index.ts",
+        output: {
+            file: "./dist/hyperapp-apollo.js",
+            format: "iife",
+            name: "HyperappApollo",
+            sourcemap: false,
+            globals: {
+                hyperapp: "hyperapp"
+            },
+            external: [
+                "hyperapp"
+            ]
+        },
+        plugins: [
+            typescript({
+                compilerOptions: {
+                    declaration: false,
+                }
+            })
+        ]
+    },
+    {
+        input: "./src/index.ts",
+        output: {
+            file: "./dist/hyperapp-apollo.min.js",
+            format: "iife",
+            name: "HyperappApollo",
+            sourcemap: false,
+            globals: {
+                hyperapp: "hyperapp"
+            },
+            external: [
+                "hyperapp"
+            ]
+        },
+        plugins: [
+            typescript({
+                compilerOptions: {
+                    declaration: false,
+                }
+            }),
+            uglify(),
         ]
     }
 ]